hushh_mcp/agents/calendar_agent/state/gemini_llm.py
import os
import google.generativeai as genai

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

if not GEMINI_API_KEY:
    raise ValueError("GEMINI_API_KEY not set in environment.")

genai.configure(api_key=GEMINI_API_KEY)
model = genai.GenerativeModel("gemini-2.5-pro")

def gemini_chat(prompt: str, system: str = None) -> str:
    """
    Sends a prompt to Gemini and returns the response.
    """
    messages = []
    if system:
        messages.append({"role": "system", "content": system})
    messages.append({"role": "user", "content": prompt})
    response = model.generate_content(messages)
    return response.text.strip()